requirements.txt
google-generativeai==0.7.2
streamlit==1.37.0
python-dotenv==1.0.1
streamlit
google-generativeai